Refactor CodeEditor
Move to overview
Change JavaScript imports in CodeEditor
Remove JavaScript from metamodel\code-editor\utils
